requirements.txt
google~=3.0.0
PyYAML~=6.0.1